Organizing components, breaking up election form component
