+++ frontend/src/App.tsx
@@ -6,9 +6,9 @@ import authSession from './authSession'
 import Header from './components/Header'
 import Elections from './components/Elections'
 import Login from './components/Login'
-import AddElection from './components/AddElection'
-import Election from './components/Election'
-import DuplicateElection from './components/DuplicateElection'
+import AddElection from './components/ElectionForm/AddElection'
+import Election from './components/Election/Election'
+import DuplicateElection from './components/ElectionForm/DuplicateElection'
 import Sandbox from './components/Sandbox'
 import DebugPage from './components/DebugPage'
