+++ frontend/src/components/Election/Admin/EditElectionRoll.tsx
@@ -5,7 +5,7 @@ import Button from "@material-ui/core/Button";
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
 import { Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Tooltip } from "@material-ui/core";
-import useFetch from "../useFetch";
+import useFetch from "../../../useFetch";
 
 const EditElectionRoll = ({ roll, onClose, fetchRolls, id}) => {
     const [updatedRoll, setUpdatedRoll] = useState(roll)
