+++ frontend/src/components/Election/Admin/ViewElectionRolls.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-import useFetch from "../useFetch";
+import useFetch from "../../../useFetch";
 import { useParams } from "react-router";
 import React from 'react'
 import Button from "@material-ui/core/Button";
