+++ frontend/src/components/Election/Results/Results.js
