+++ frontend/src/components/Election/Sidebar.tsx
