+++ frontend/src/components/Election/VoterAuth.tsx
