+++ frontend/src/components/Election/Voting/StarBallot.js
