+++ frontend/src/components/ElectionForm/AddElection.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import useFetch from '../useFetch';
+import useFetch from '../../useFetch';
 import ElectionForm from "./ElectionForm";
 import Container from '@material-ui/core/Container';
 import { useNavigate } from "react-router"
