+++ frontend/src/components/ElectionForm/DuplicateElection.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useParams } from "react-router";
-import useFetch from "../useFetch";
+import useFetch from "../../useFetch";
 import Container from '@material-ui/core/Container';
 import ElectionForm from "./ElectionForm";
 import { useState, useEffect } from "react";
